Move tombstones to their own thing.
